♻️  zu: Split `signature::Error` into a separate mod/file
