+++ zvariant_utils/src/signature/error.rs
@@ -1,0 +1,16 @@
+use core::fmt;
+
+/// Enum representing the max depth exceeded error.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum Error {
+    /// Invalid signature.
+    InvalidSignature,
+}
+
+impl fmt::Display for Error {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::InvalidSignature => write!(f, "Invalid signature"),
+        }
+    }
+}
